backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
nltk==3.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
